+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## v0.3.0-5
 
 - Fixed the minimal SSL/TLS protocol version restriction in virtual hosts.
+- Added the router for controllers' methods mapping.
 
 ## v0.3.0-4